+++ server/api/blog/model/request/menu.go
@@ -14,7 +14,7 @@ package request
 //	FrameLoading bool        `json:"frameLoading,omitempty"` // iframe页是否开启首次加载动画
 //	Transition   Transition  `json:"transition,omitempty"`   // 页面加载动画
 //	HiddenTag    bool        `json:"hiddenTag,omitempty"`    // 是否不添加信息到标签页
-//	DynamicLevel int         `json:"dynamicLevel,omitempty"` // 动态路由可打开的最大数量
+//	DynamicLevel int64         `json:"dynamicLevel,omitempty"` // 动态路由可打开的最大数量
 //	ActivePath   string      `json:"activePath,omitempty"`   // 将某个菜单激活
 //}
 
@@ -23,7 +23,7 @@ type Meta struct {
 	Title    string `json:"title"`              // 菜单名称
 	Icon     string `json:"icon,omitempty"`     // 菜单图标
 	ShowLink bool   `json:"showLink,omitempty"` // 是否在菜单中显示
-	Rank     int    `json:"rank,omitempty"`     // 菜单升序排序
+	Rank     int64  `json:"rank,omitempty"`     // 菜单升序排序
 
 	// 子菜单才有的属性
 	ExtraIcon    interface{} `json:"extraIcon,omitempty"`    // 菜单名称右侧的额外图标
@@ -35,7 +35,7 @@ type Meta struct {
 	FrameLoading bool        `json:"frameLoading,omitempty"` // iframe页是否开启首次加载动画
 	Transition   Transition  `json:"transition,omitempty"`   // 页面加载动画
 	HiddenTag    bool        `json:"hiddenTag,omitempty"`    // 是否不添加信息到标签页
-	DynamicLevel int         `json:"dynamicLevel,omitempty"` // 动态路由可打开的最大数量
+	DynamicLevel int64       `json:"dynamicLevel,omitempty"` // 动态路由可打开的最大数量
 	ActivePath   string      `json:"activePath,omitempty"`   // 将某个菜单激活
 }
 
@@ -58,7 +58,7 @@ type Transition struct {
 
 // RouteConfigsTable 对应TypeScript中的RouteConfigsTable接口
 type RouteConfigsTable struct {
-	Type      int                 `json:"type"`                // 菜单类型（0代表菜单、1代表iframe、2代表外链、3代表按钮）
+	Type      int64               `json:"type"`                // 菜单类型（0代表菜单、1代表iframe、2代表外链、3代表按钮）
 	Path      string              `json:"path"`                // 路由地址
 	Name      string              `json:"name,omitempty"`      // 路由名字
 	Component interface{}         `json:"component,omitempty"` // Layout组件
